+++ src/main/java/com/webapp/site/entities/Monument.java
@@ -1,7 +1,7 @@
 package com.webapp.site.entities;
 
 import java.io.Serializable;
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 
 /**
@@ -16,7 +16,7 @@ public class Monument implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int idArt;
+	private Long idArt;
 
 	//bi-directional one-to-one association to Art
 	@OneToOne
@@ -31,11 +31,11 @@ public class Monument implements Serializable {
 	public Monument() {
 	}
 
-	public int getIdArt() {
+	public Long getIdArt() {
 		return this.idArt;
 	}
 
-	public void setIdArt(int idArt) {
+	public void setIdArt(Long idArt) {
 		this.idArt = idArt;
 	}
