+++ src/main/java/com/webapp/site/entities/Painting.java
@@ -1,7 +1,7 @@
 package com.webapp.site.entities;
 
 import java.io.Serializable;
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 
 /**
@@ -16,7 +16,7 @@ public class Painting implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int idArt;
+	private Long idArt;
 
 	//bi-directional one-to-one association to Art
 	@OneToOne
@@ -36,11 +36,11 @@ public class Painting implements Serializable {
 	public Painting() {
 	}
 
-	public int getIdArt() {
+	public Long getIdArt() {
 		return this.idArt;
 	}
 
-	public void setIdArt(int idArt) {
+	public void setIdArt(Long idArt) {
 		this.idArt = idArt;
 	}
